Add files via upload
The following code uses an asymmetric walk to generate L paths of length N.
This code looks at two the cases of one-boundary levels and two-boundary level.
In the case of one boundary reflection only occurs once and several plots of
m = 0.5, 1, 2 are shown. In the case of two boundary reflection occurs each time
the path tries to cross one of the boundaries. 5 plot at a = -1 and b = 1 are 
superimposed n one plot. Expectation of reaching the boundary for the first time 
and the probability of the path being within the boundary in either 
case are calculated at N = 1000 and L = 1000. The expected number of reflection is 
determined for the two boundary case.